feat: implement sign up / in buttons
